ID-10 Add prettier and eslint
