+++ package.json
@@ -7,6 +7,9 @@
     "prebuild": "rimraf ./dist",
     "build": "babel ./src -d ./dist",
     "prepublish": "npm run build",
+    "lint": "eslint --ext .js src/",
+    "lint:fix": "eslint --ext .js src/ --fix",
+    "pretest": "npm run lint",
     "test": "mocha --require @babel/register ./src/**/*.test.js"
   },
   "repository": {
@@ -36,8 +39,16 @@
     "@babel/plugin-proposal-class-properties": "^7.7.4",
     "@babel/preset-env": "^7.7.6",
     "@babel/register": "^7.7.4",
+    "babel-eslint": "^10.1.0",
     "chai": "^4.2.0",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-config-prettier": "^6.10.0",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-prettier": "^3.1.2",
     "mocha": "^6.2.2",
+    "prettier": "1.19.1",
     "rimraf": "^3.0.2",
     "sinon": "^7.5.0"
   }
